+++ src/splat/segtypes/common/data.py
@@ -60,11 +60,7 @@ class CommonSegData(CommonSegCodeSubsegment, CommonSegGroup):
             if preamble:
                 f.write(preamble + "\n")
 
-            f.write(f".section {self.get_linker_section()}")
-            section_flags = self.get_section_flags()
-            if section_flags:
-                f.write(f', "{section_flags}"')
-            f.write("\n\n")
+            f.write(f"{self.get_section_asm_line()}\n\n")
 
             f.write(self.spim_section.disassemble())