+++ src/splat/segtypes/psx/asm.py
@@ -4,8 +4,7 @@ from ..common.asm import CommonSegAsm
 
 
 class PsxSegAsm(CommonSegAsm):
-    @staticmethod
-    def get_file_header():
+    def get_file_header(self):
         ret = []
 
         ret.append('.include "macro.inc"')
@@ -17,7 +16,8 @@ class PsxSegAsm(CommonSegAsm):
         if preamble:
             ret.append(preamble)
             ret.append("")
-        ret.append('.section .text, "ax"')
+
+        ret.append(self.get_section_asm_line())
         ret.append("")
 
         return ret